pyproject.toml
[tool.poetry]
name = "blum-app"
version = "0.1.0"
description = "Bot for automatization tasks in blum app"
authors = ["Kazantsev Artem <me@lokach.ru>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
loguru = "^0.7.3"
fake-headers = "^1.0.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.0.1"
black = "^24.10.0"
ruff = "^0.8.3"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 89
include = '\.pyi?$'

[tool.ruff]
lint.select = ["F", "E", "W", "I", "N", "PL", "C", "B"]
line-length = 89

[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]
